oif-bt/requirements.txt
fastapi==0.95.0
motor==3.1.1
paho-mqtt==1.6.1
toml==0.10.2
tomli==2.0.1
uvicorn==0.21.1
